% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/zzz.R
\name{getCores}
\alias{getCores}
\title{Returns the number of cores.}
\usage{
getCores(cores)
}
\arguments{
\item{cores}{the number of cores, it is 1 in windows platform.}
}
\description{
Returns the number of cores.
}
\examples{
cores <- getCores(2L)
}

