% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/limitToXkb.R
\name{readBed}
\alias{readBed}
\title{readBed Returns a GenomicRanges object constrcuted from the specified bed
file.}
\usage{
readBed(file, ...)
}
\arguments{
\item{file}{Path to the input file.}

\item{...}{Extra argument passed to read.table}
}
\value{
Returns GRanges object representing mapped reads.
}
\description{
Bed file format is assumed to be either four column: seqnames, start, end,
strand columns; or six column: seqnames, start, end, name, score, and strand.
Three column format is also possible when there is no strand information.
}
\details{
Any additional arguments availiable to read.table can be specified.
}
\examples{
fileBed <- system.file(package = "rtracklayer", "tests", "test.bed")
readBed(fileBed, skip = 2, colClasses = c(rep(NA, 6), rep("NULL", 6)))
}
\author{
Minho Chae and Charles G. Danko.
}

