% Generated by roxygen2 (4.0.1): do not edit by hand
\name{countMappableReadsInInterval}
\alias{countMappableReadsInInterval}
\title{countMappableReadsInInterval counts the number of mappable reads in a set
of genomic features.}
\usage{
countMappableReadsInInterval(features, UnMap, debug = FALSE, ...)
}
\arguments{
\item{features}{A GRanges object representing a set of genomic coordinates.
The meta-plot will be centered on the start position.}

\item{UnMap}{List object representing the position of un-mappable reads.
Default: not used.}

\item{debug}{If set to TRUE, provides additional print options.
Default: FALSE}

\item{...}{Extra argument passed to mclapply}
}
\value{
Returns a vector of counts, each representing the number of reads
inside each genomic interval.
}
\description{
Supports parallel processing using mclapply in the 'parallel' package.
To change the number of processors, use the argument 'mc.cores'.
}
\author{
Charles G. Danko and Minho Chae
}

