% Generated by roxygen2 (4.0.1): do not edit by hand
\name{evaluateHMMInAnnotations}
\alias{evaluateHMMInAnnotations}
\title{evaluateHMM Evaluates HMM calling.}
\usage{
evaluateHMMInAnnotations(tx, annox)
}
\arguments{
\item{tx}{GRanges of transcripts predicted by HMM.}

\item{annox}{GRanges of non-overlapping annotatoins.}
}
\value{
a list of error information; merged annotations, dissociated annotation,
total, and rate.
}
\description{
Evaluates HMM calling of transripts compared to known annotations.
}
\examples{
tx <- GRanges("chr7", IRanges(start=seq(100, 1000, by=200),
width=seq(100, 1000, by=100)), strand="+")
annox <- GRanges("chr7", IRanges(start=seq(110, 1100, by=150),
width=seq(100, 1000, by=150)), strand="+")
error <- evaluateHMMInAnnotations(tx, annox)
}
\author{
Minho Chae
}

