% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressedGenes.R
\name{expressedGenes}
\alias{expressedGenes}
\title{Function identifies expressed features using the methods introduced in
Core, Waterfall, Lis; Science, Dec. 2008.}
\usage{
expressedGenes(features, reads, Lambda = NULL, BPPARAM = bpparam())
}
\arguments{
\item{features}{A GRanges object representing a set of genomic coordinates.
The meta-plot will be centered on the start position.}

\item{reads}{A GRanges object representing a set of mapped reads.}

\item{Lambda}{Measurement of assay noise.  Default: 0.04 reads/ kb in a
library of 10,751,533 mapped reads. (background computed in Core,
Waterfall, Lis. (2008) Science.).}

\item{BPPARAM}{Registered backend for BiocParallel.
Default: BiocParallel::bpparam()}
}
\value{
Returns a data.frame representing the expression p.values for
features of interest.
}
\description{
Function identifies expressed features using the methods introduced in
Core, Waterfall, Lis; Science, Dec. 2008.
}
\author{
Charles G. Danko
}
