% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTxDensity.R
\name{getTxDensity}
\alias{getTxDensity}
\title{getTxDensity Calculates transcript density.}
\usage{
getTxDensity(tx, annox, plot = FALSE, scale = 1000L, ...)
}
\arguments{
\item{tx}{GRanges of transcripts.}

\item{annox}{GRanges of non-overlapping annotations.}

\item{plot}{Logical.  If TRUE, plot transcript density.  Default: FALSE}

\item{scale}{Numeric. Scaled size of a gene for transcript density
calculation.
Default: 1000L}

\item{...}{Extra argument passed to mclapply.}
}
\value{
Returns a list of FTD, TTD, PostTTS, and TUA.
}
\description{
Calculates transcript density for transcripts which overlaps with
annotations.
For 'combined' or 'broken up' errors,
best overlapped transcripts or annotations are used.
}
\details{
Supports parallel processing using mclapply in the 'parallel' package.
To change the number of processors
set the option 'mc.cores'.
}
\examples{
library(GenomicRanges)
tx <- GRanges("chr7", IRanges(start=seq(1000,4000, by=1000),
width=seq(1000, 1300, by=100)), strand=rep("+", 4))
annox <- GRanges("chr7", IRanges(start=seq(1100,4100, by=1000),
width=seq(900, 1200, by=100)), strand=rep("+", 4))
## Not run:
# density <- getTxDensity(tx, annox)
}
\author{
Minho Chae
}
