% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeConsensusAnnotations.R
\name{makeConsensusAnnotations}
\alias{makeConsensusAnnotations}
\title{makeConsensusAnnotations Makes a consensus annotation}
\usage{
makeConsensusAnnotations(ar, minGap = 1L, minWidth = 1000L,
  column = "gene_id", BPPARAM = bpparam())
}
\arguments{
\item{ar}{GRanges of annotations to be collapsed.}

\item{minGap}{Minimum gap between overlapped annotations after truncated.
Default: 1L}

\item{minWidth}{Minimum width of consensus annotations. Default: 1000L}

\item{column}{Column by which to group transcripts.}

\item{BPPARAM}{Registered backend for BiocParallel.
Default: BiocParallel::bpparam()}
}
\value{
Returns GRanges object of annotations.
}
\description{
Makes a non-overlapping consensus annotation.  Gene annotations are often
overlapping due to #' multiple isoforms for a gene.
In consensus annotation, isoforms are first reduced so that only
redundant intervals are used to represent a genomic interval for a gene,
i.e., a gene id.
Remaining unresolved annotations are further reduced by truncating 3'
end of annotations.
}
\examples{
library(TxDb.Hsapiens.UCSC.hg19.knownGene)
tx <- transcripts(
    TxDb.Hsapiens.UCSC.hg19.knownGene,
    columns=c("gene_id", "tx_id", "tx_name"),
    filter=list(tx_chrom="chr7"))
ca <- makeConsensusAnnotations(tx)
}
\author{
Minho Chae
}
