% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pausingIndex.R
\name{pausingIndex}
\alias{pausingIndex}
\title{Returns the pausing index for different genes.  TODO: DESCRIBE THE PAUSING
INDEX.}
\usage{
pausingIndex(features, reads, size = 50, up = 1000, down = 1000,
  BPPARAM = bpparam())
}
\arguments{
\item{features}{A GRanges object representing a set of genomic coordinates.}

\item{reads}{A GRanges object representing a set of mapped reads.}

\item{size}{The size of the moving window.}

\item{up}{Distance upstream of each f to align and histogram.}

\item{down}{Distance downstream of each f to align and histogram (NULL).}

\item{BPPARAM}{Registered backend for BiocParallel.
Default: BiocParallel::bpparam()}
}
\value{
Returns a data.frame of the pausing indices for the input genes.

Returns the pausing index for different genes.
}
\description{
Returns the pausing index for different genes.  TODO: DESCRIBE THE PAUSING
INDEX.
}
\examples{
library(GenomicAlignments)
features <- GRanges("chr7", IRanges(2394474,2420377), strand="+")
reads <- as(readGAlignments(system.file("extdata", "S0mR1.bam",
            package="groHMM")), "GRanges")
## Not run:
# pi <- pausingIndex(features, reads)
}
\author{
Charles G. Danko and Minho Chae.
}
