% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pausingIndex.R
\name{pausingIndex}
\alias{pausingIndex}
\title{Returns the pausing index for different genes.  TODO: DESCRIBE THE PAUSING 
INDEX.}
\usage{
pausingIndex(features, reads, size = 50, up = 1000, down = 1000,
  UnMAQ = NULL, debug = FALSE, ...)
}
\description{
Supports parallel processing using mclapply in the 'parallel' package.  
To change the number of processors, use the argument 'mc.cores'.
}
\details{
@param features A GRanges object representing a set of genomic coordinates.
 @param reads A GRanges object representing a set of mapped reads.
 @param size The size of the moving window.
 @param up Distance upstream of each f to align and histogram.
 @param down Distance downstream of each f to align and histogram (NULL).
 @param UnMAQ Data structure representing the coordinates of all un-mappable
 regions in the genome.
 @param debug If set to TRUE, provides additional print options. 
 Default: FALSE
 @param ... Extra argument passed to mclapply
 @return Returns a data.frame of the pausing indices for the input genes.
 @author Charles G. Danko and Minho Chae.
 @return Returns the pausing index for different genes.
 @examples
 features <- GRanges("chr7", IRanges(2394474,2420377), strand="+") 
 reads <- as(readGAlignments(system.file("extdata", "S0mR1.bam", 
             package="groHMM")), "GRanges")
 ## Not run:
 # pi <- pausingIndex(features, reads)
}
