% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Rnorm}
\alias{Rnorm}
\title{Rnorm fits a normal distribution to a specified data vector using maximum
likelihood.}
\usage{
Rnorm(X)
}
\arguments{
\item{X}{A vector of observations, assumed to be real numbers in the
inveraval (-Inf,+Inf).}
}
\value{
Returns a list of parameters for the best-fit normal distribution
(mean and varience).
}
\description{
Rnorm fits a normal distribution to a specified data vector using maximum
likelihood.
}
\author{
Charles G. Danko
}

