% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlsFits.R
\name{tlsSvd}
\alias{tlsSvd}
\title{A 'total least squares' implementation using singular value demposition.}
\usage{
tlsSvd(x, y)
}
\arguments{
\item{x}{X values.}

\item{y}{Y values.}
}
\value{
Parameters for the linear model Y~a*X+e.
}
\description{
A 'total least squares' implementation using singular value demposition.
}
\author{
Charles G. Danko
}
