% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windowAnalysis.R
\name{windowAnalysis}
\alias{windowAnalysis}
\title{windowAnalysis Returns a vector of integers representing the counts of 
reads in a moving window.}
\usage{
windowAnalysis(reads, strand = "*", windowSize = stepSize,
  stepSize = windowSize, chrom = NULL)
}
\arguments{
\item{reads}{GenomicRanges object representing the position of reads 
mapping in the genome.}

\item{strand}{Takes values of "+", "-", or "*".  "*" denotes collapsing 
reads on both strands.  Default: "*".}

\item{windowSize}{Size of the moving window. Either windowSize or 
stepSize must be specified.}

\item{stepSize}{The number of bp moved with each step.}

\item{chrom}{Chromosome for which to return data.
Default: returns all avaliable data.}
}
\value{
Returns a list object, each element of which represents a 
chromosome.
}
\description{
windowAnalysis Returns a vector of integers representing the counts of 
reads in a moving window.
}
\examples{
S0mR1 <- as(readGAlignments(system.file("extdata", "S0mR1.bam",
     package="groHMM")), "GRanges")
## Not run:
# Fp <- windowAnalysis(S0mR1, strand="+", windowSize=50)
}
\author{
Charles G. Danko and Minho Chae
}
